Read coord system status before motor positions
Read the CS (coordinate system) status before reading the motor
positions so that the CS in-position status is consistent with the
motor positions.  The CS in-position status is used to set the
"motorAxisMoving" and "motorAxisDone" parameters of the Asyn motors in
the CS.  If the CS status is read *after* the motor positions are read,
there's a small window of time where the motor position of an Asyn motor
could be *not* in-position (i.e., the position would not be considered
in-position by the PMAC controller), but the done-moving flag of the
Asyn motor could be true.  To avoid this, read the CS status *before*
reading the motor positions.